backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
biopython==1.83
numpy==1.26.4
pydantic==2.7.1
python-multipart==0.0.9
